Projects/Batch-2022-2026/net-gaurd/net-gaurd/requirements.txt
Flask==3.0.0
APScheduler==3.10.4
python-dotenv==1.0.0
